% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kaya_Identity.R
\name{Kaya_Equation}
\alias{Kaya_Equation}
\title{Illustration of Kaya_Equation}
\usage{
Kaya_Equation(pop, gdp, enInt, carbInt)
}
\arguments{
\item{pop}{: population size (in millions)}

\item{carbInt}{Carbon Intensity (in tonnes CO2/Gigajoule).
Carbon Intensity is the CO2 emitted for produced energy.
This number depends on the energy mix used (coal, solar, . . . ).}

\item{gdp:}{GDP per capita (in 1000$/person)}

\item{enInt:}{Energy Intensity (in Gigajoule/$1000GDP). Energy Intensity is t
he energy needed to produce a certainamount of economic value.}
}
\value{
numeric number co2 as Kaya Identity
examples Kaya_Equation(pop=82.4,gdp=44,enInt = 5,carbInt = 0.05)
}
\description{
Illustration of Kaya_Equation
}
